requirements.txt
-e .
channels_redis>=3.4,<4.0
django-cors-middleware
django-extensions
django-debug-toolbar
mysqlclient
